requirements.txt
pandas==1.3.3
Pillow==8.4.0
scikit_learn==1.0.1
streamlit==0.62.0
